feat(db) query semaphore for Postgres connector
This commit introduces an optional semaphore locking mechanism
within the Postgres connector :query() method. This provides a solution
to the longstanding problem of OpenResty not providing a ceiling
on the number of connections that can be established to a remote
TCP host. As Postgres is sensitive about the number of connections opened
to a given server, this provides a workaround solution until a native
connection cap implementation is made available for OpenResty.

This commit introduce two new config options, 'pg_semaphore_max' and
'pg_semaphore_timeout'. The former defines the number of resources
allocated to a `ngx.semaphore` instance that is bound to a Postgres
connector object. This effectively defines the maximum number of
concurrent queries to a Postgres server that can be issued from a
given worker process (the more natural number of maximum number of
queries issued by a given Kong node can then be extrapolated by multiplying
this value by the number of Kong worker processes - using the native
ngx semaphore mechanism seems the more appropriate solution vs. creating
a node-wide semaphore (via some messy shared dict approach) that provided
a slightly more human-natural accounting).

The default 'pg_semaphore_max' value of 0 results in no semaphore object
created in the connector, and no lock acquisition performed before
executing a query. This leaves the existing behavior in place.